Make the new-base flag default to True
When building `curl` with `--allow-newer=base` (which is often the case when I'm testing packages with GHC HEAD), the `new-base` flag has a tendency to be set to `False`, resulting in build failures (see https://github.com/haskell/cabal/issues/3457#issuecomment-221487085 for further discussion). A simple workaround is to make `new-base` default to `True`, which avoids the issue.